thread safety
the callback closures are sent across a thread boundary at least once,
making it nessesary to mark them as Send.
I'm not sure if marking them as Sync is strictly nessesary, as each callback is only really executed on one thread at once,
but since that's dependend on the behavior of an undocumented API
I just threw it in there for extra safety.

I also tried to simplify the cast in callback(), with fewer raw pointers and less mut
(which is likely evil in this context anyways)